requirement.txt
Pillow==12.1.0
numpy==1.26.4
tensorflow==2.16.1
tf-keras==2.16.0
opencv-python==4.11.0.86
deepface==0.0.93
ultralytics
matplotlib
pygame
pandas


#pour lancer l'installation, pip install -r requirement.txt